requirements.txt
requests>=2.27.1
google-cloud-bigquery>=2.33.0
google-cloud-bigquery-storage>=2.11.0
pandas-gbq==0.16.0; python_version > '3.6'
pyarrow==6.0.1
pandas==1.4.1
google-auth>=2.16.2
simple-salesforce==1.12.5